third_party/blink/web_tests/platform/generic/fast/forms/textarea/rows-attribute-expected.txt

HTMLTextAreaElement rows attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement("textarea").rows is 2
PASS rowsAttributeEffect("") is 2
PASS rowsAttributeEffect("1") is 1
PASS rowsAttributeEffect("2") is 2
PASS rowsAttributeEffect("10") is 10
PASS rowsAttributeEffect("0") is 2
PASS rowsAttributeEffect("-1") is 2
PASS rowsAttributeEffect("1x") is 1
PASS rowsAttributeEffect("1.") is 1
PASS rowsAttributeEffect("1.9") is 1
PASS rowsAttributeEffect("2x") is 2
PASS rowsAttributeEffect("2.") is 2
PASS rowsAttributeEffect("2.9") is 2
PASS rowsAttributeEffect("a") is 2
PASS rowsAttributeEffect("7") is 2
PASS rowsAttributeEffect("  7") is 7
PASS rowsAttributeEffect(arabicIndicDigitOne) is 2
PASS rowsAttributeEffect("2" + arabicIndicDigitOne) is 2
PASS rowsAttributeEffect("2147483646") is 2147483646
PASS rowsAttributeEffect("2147483647") is 2147483647
PASS rowsAttributeEffect("2147483648") is 2
PASS rowsAttributeEffect("2147483649") is 2
PASS rowsAttributeEffect("4294967295") is 2
PASS rowsAttributeEffect("4294967296") is 2
PASS rowsAttributeEffect("4294967297") is 2
PASS successfullyParsed is true

TEST COMPLETE
